+++ src/jackhmmer.c
@@ -534,7 +534,7 @@ serial_master(ESL_GETOPTS *go, struct cfg_s *cfg)
 #endif
 
   infocnt = (ncpus == 0) ? 1 : ncpus;
-  ESL_ALLOC(info, sizeof(*info) * infocnt);
+  ESL_ALLOC(info, (ptrdiff_t) sizeof(*info) * infocnt);
 
   /* Ready to begin */
   output_header(ofp, go, cfg->qfile, cfg->dbfile);
